Move wallet panel page handler into its own dir
To track deps better
