+++ browser/ui/webui/wallet_panel/page_handler/BUILD.gn
@@ -1,0 +1,17 @@
+# Copyright (c) 2021 The Brave Authors. All rights reserved.
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at http://mozilla.org/MPL/2.0/.
+
+source_set("page_handler") {
+  sources = [
+    "wallet_panel_page_handler.cc",
+    "wallet_panel_page_handler.h",
+  ]
+  deps = [
+    "//brave/components/brave_wallet_ui:mojo_bindings",
+    "//content/public/browser",
+    "//mojo/public/cpp/bindings",
+    "//ui/webui:webui",
+  ]
+}
