+++ browser/ui/webui/wallet_panel/page_handler/wallet_panel_page_handler.cc
@@ -3,7 +3,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // you can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "brave/browser/ui/webui/wallet_panel/wallet_panel_page_handler.h"
+#include "brave/browser/ui/webui/wallet_panel/page_handler/wallet_panel_page_handler.h"
 
 #include <utility>
