+++ browser/ui/webui/wallet_panel/page_handler/wallet_panel_page_handler.h
